Avoid infinite recursion when exiting
When exiting the game due to death, the database was closed
prior to calling exit_() which resulted in failure to be able
to report the score properly, which resulted in calling exit_().

This checks if the database is closed and skips double-reporting
the score if the earlier code has already closed it.